CI on PRs, fix build warning, fixed rust version
This commit updates CI to also run on PRs. Additionally CI now fails if
build warnings are encountered.

A `rust-toolchain.toml` file has been added to the root of the repo that
ensures that all collaboraters use the same up-to-date Rust version.
Cargo picks up this file automatically and uses rustup to install and
use the right version of rustc and cargo for each project.

CI now also adheres to this toolchain file through the use of
`dtolnay/rust-toolchain@stable`.