Add `resolve()` method to `SourceProvider` (#177)
* Refactor bundle tests' `fs!` macro to drop dependency on `TestProvider` (#174)

This allows the `fs!` macro to be reused with other `SourceProvider` implementations on a per-test basis.

* Refactor `error_test()` to invoke a callback for its assertion (#174)

This allows each caller to define its own assertions and not be forced into asserting for `BundleErrorKind::UnsupportedLayerCombination`.

* Add `resolve()` method to `SourceProvider` (#174)

This adds a `resolve()` method to `SourceProvider` which is responsible for converting an `@import` specifier to a file path. The default `FileProvider` still uses the existing behavior of assuming the specifier to be a relative path and joining it with the originating file's path.