Add `JsSourceProvider` and use it in `bundleAsync()` (#174)
This adds a new `JsSourceProvider` which acts a `SourceProvider` which invokes associated JS implementations of the resolver if they exist or falls back to `FileProvider` when not given. This allows JS consumers to override one or both of these methods. If JS does *not* override either, they should not pay any significant performance penalty since all the Rust work will stay on the same thread.

The JS implementations are invoked as thread-safe functions, pushing the arguments onto a queue and adding it to the event queue. Some time later, the main thread pulls from this queue and invokes the function. `napi-rs` doesn't seem to provide any means of receiving a JS return value in Rust, so instead arguments for both `read()` and `resolve()` include a callback function using Node calling conventions (`callback(err, result)`). This looks like:

```javascript
await bundleAsync({
  // Options...

  resolver: {
    read(file: string, cb: (string) => void): void {
      // Read file and invoke callback with result.
      fs.read(file, 'utf-8').then((res) => cb(null, res), (err) => cb(err, null));
    },

    resolve(specifier: string, originatingFile: string, cb: (string) => void): void {
      // Resolve and invoke callback with result.
      const resolved = path.resolve(originatingFile, '..', specifier);
      cb(null, resolved);
    },
  },
});
```