Add JS shim and types for `bundleAsync()` so the public API is just `(...args) => Promise<Result>` (#174)
This hides the fact that `napi-rs` doesn't actually do anything with the return value of a threadsafe JS function, so any returned data would be dropped. Instead, Parcel adds a callback function which gets invoked by JS with the resulting value, following Node conventions of `callback(err, result)`. This is unergonimic as an API, so the JS shim exposes a `Promise`-based interface which gets converted to the callback behavior required by `napi-rs`.

This also includes tests for all the common use cases and error conditions of custom JS resolvers.