Fix calc in @media
Fixes #195.

Adds a new AddInternal trait used for performing addition within a calc(). This doesn't wrap sums in calc() functions, because that's not necessary for nested sums within math functions. The result is wrapped in calc() if necessary during parsing. Later, if the calc() needs to be modified, we need to unwrap this calc, perform the addition, and re-wrap if necessary. That's now done by the public std::ops::Add implementation.