fix: resolution = 1 cannot be omitted from `image-set()`
Before this PR, we would always remove it when resolution = `1x/1dppx`. This is not compliant with the specification and will not render in the browser.

Input:

```css
body {
  background-image: image-set("../images/bg.webp" 1x, "../images/bg.webp" 2x);
}
```

Before:

```css
body {
  background-image: image-set("../images/bg.webp", "../images/bg.webp" 2x);
}
```

After:

```css
body {
  background-image: image-set("../images/bg.webp" 1x, "../images/bg.webp" 2x);
}
```

In the future we need to add detection of resolution <= 0.
For example: `image-set(url(example.png) 0x)` is an invalid syntax.