include whitespace when parsing the `syntax: '*'` values
I know that this isn't 100% correct because we _only_ want to do that
for the `initial-value` property and not for everything.

We also "abuse" the fact that `initial-value:;` results in an
`EndOfInput` error and then we conver that to `Whitespace("")`.

But the idea is as follows:
Instead of skipping whitespace, allow it and even allow "nothing". If we
skip whitespace, the current parser expects a valid token but there is
none and therefore results in an `EndOfInput`.

But in case of `initial-value: ;` or `initial-value:;` these values are
allowed.

Later, we can "ignore" all the whitespace, and only emit a single one
(or nothing) for the `initial-value`.