move custom `initial-value` parsing to Property parser
This is the better spot because now it is specific to the
`initial-value` of an `@property` with Universal syntax.

Used `is_exhausted` to know that we are "done" or not, it will skip
whitespace so it will be true in both the `initial-value:;` and
`initial-value: ;` case.

Then we still use `parser.next_including_whitespace()` because we are
still interested in the whitespace if there is some (in case of
`initial-value: ;`). If there is no whitespace, then it results in an
end of input error which we can convert to a `WhiteSpace("")` token.

We do capture the error to default to an empty whitespace token, but we
can capture _all_ errors safely because if _something_ was actually
wrong, the wrapping `parser.is_exhausted()` would not have been ok and
therefore false.