drop prefixed properties, if non-prefix property exists
If you have the following CSS:
```css
.foo {
  transition-property: backdrop-filter;
}
```

Then the `backdrop-filter` will be prefixed such that it looks like
this:
```css
.foo {
  transition-property: -webkit-backdrop-filter, backdrop-filter;
}
```

However, if you already have `-webkit-backdrop-filter` in the list:
```css
.foo {
  transition-property: -webkit-backdrop-filter, backdrop-filter;
}
```

Then it compiles to:
```css
.foo {
  transition-property: -webkit-backdrop-filter, -webkit-backdrop-filter, backdrop-filter;
}
```

This is not what we want. So in this case, whenever a non-prefixed
property is found in the list (`backdrop-filter`), then we can remove or
ignore all prefixed values. Once we are printing the CSS back, the
`backdrop-filter` will print the prefixed version as well.

```css
.foo {
  transition-property: -webkit-backdrop-filter, backdrop-filter;
}
```